pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "openai-agents[litellm,voice]>=0.0.15",
    "pydub>=0.25.1",
    "python-socks>=2.7.1",
    "sherpa-onnx>=1.12.0",
    "sounddevice>=0.5.2",
]
